+++ packages/ol/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @summeruse/ol
 
+## 0.0.1
+
+### Patch Changes
+
+- 8e34a3f: 添加 useDrawLineString useDrawPolygon
+- b6b95dc: 添加创建几何要素的实用函数
+- 58c1e99: 添加坐标系相关函数的注释
+- 474d537: 添加常数模块
+- 474d537: 添加工具函数模块
+- 1d0394b: 添加创建样式的工具函数
+- 4739812: 添加创建矢量图层的函数
+- c0a970d: 添加`vite-plugin-css-injected-by-js`插件以将 CSS 注入 JS 文件
+- 474d537: 添加 ol-map 组件
+- 474d537: 添加 useGraticule
+- 474d537: 添加 useSwitchBaseLayer
+- 048ca5d: 修复：天地图类型
+- a0c76b5: style(n-ol-pointermove): 移除冗余的样式并改用 theme-overrides
+- 062b2f0: 添加 useDrawLineString
+- 474d537: 添加 n-ol-pointermove
+
 ## 0.0.1-beta.5
 
 ### Patch Changes